Simplified relationships, removing associated type "Data" from `MarketItem`.
`MarketItem`s no longer know about their representative data, and do not have
an initializer that takes that type. The responsibility for creating the
"Item" is entirely within the `MarketItemData` type: it acts as the factory.

To that end, the full memberwise initializers for the existing `MarketItem` types, `Purchase` and `Merch`, have been moved to private extensions in their corresponding "Data" class files.

The impetus here was both to remove the knowledge of "Data" types from "Item"
types and also to simplify `MarketListData`'s generic parameters, which are
no longer so gnarly.