Added root Archving class `MarketRepository` with tests.
`MarketRepository` knows how to transform a `Market` to `MarketData` and
write that Data using an `NSCoder` archiver. It can also use an
`NSCoder` unarchiver to create a `Market`, given a key -- the `Market`'s
`UniqueID` -- present in the archive.

The repository throws an error if asked to read or write without the
appropriate `NSCoder` instance. It also throws an error if asked to read a
key that does not exist in the provided archive.