Added tests for all `Purchase` functionality.
In the process, the interface changed slightly. Rather than
`changingMerch(to:)`, a `Purchase` now allows changing its name and unit
individually. It internally handles the necessary creation of the new `Merch`.

The type of `quantity` changed to be non-optional. This was changed to prevent having to map from 0 to `nil` in multiple places. The "zero quantity equals nil" test was removed.