Converted `MerchData` to use Swift `Encodable`, `Decoder`, and `Encoder`.
These are adapters for `NSCoder`/`NSCoding`, giving a little bit more type
information and allowing for future use of archivers that are not based on
`NSCoder`.

`NSCoding` conformance was removed from `MarketItemData` in the course of
this change, and added to `PurchaseData`. After `PurchaseData` is converted,
the conformance to `Encodable` will be moved back up to the protocol.