Converted `MarketListData` and `MarketData` to use Encodable/Encoder/Decoder.
Changed the signature of `encode(encodable:forKey:)` and ; the "encodables"
are now `AnyObject` rather than `Encodable`/`Decodable`. This sprang from
problems implementing `decodeArray(forKey:)`. It's not possible to cast an
array to a non-ObjC protocol like `Encodable`, so the return type of
`decodeArray(forKey:)` has no apparent choice but to be `[AnyObject]?`.

A good future option would be to create a type eraser type for `Encodable`.