Refactor some Merch creation in tests
Added three properties to Merch+Dummy, `dummyNames`, `dummies`, and
`offListDummy`. The first is a list of names to use for creating test `Merch`
instances. The second uses that list to create a list of test `Merch`es. The
third is a `Merch` instance that it not on the list, for testing search and
update of `MarketList`.

Added a `dummy` static property to `Purchase` as well for testing: it creates
a `Purchase` instance using `Merch.dummy`.

Added missing `Purchase` test, for `quantity` mutation preserving other
fields. Improved `Purchase` mutation-preservation tests to explicitly
initialize all fields of the tested instance, rather than using the defaults.