Tests for merch and purchase presentations
Added tests for their exposed field values and for their notification
behavior. They both handle input from their view and respond with a
`didUpdate` callback to prompt the view to read the new values. A parent
presentation or other interested observer is also notified with more specific
callbacks, passing old and/or new values when appropriate.

`MerchPresentation` and `PurchasePresentation` are currently added to
LoafBottleStickKit in the Xcode project rather than to the app target. This
will be reshuffled later as needed.

Removed `isInPurchase` from `MerchPresentation` because it needs to be
redesigned and was not tested.