Tests for `InventoryPresentation`
Basic functionality for `InventoryPresentation` is established. The class
exposes a sorted list of `MerchPresentation`s, and accepts additions of
`Merch` and deletion by index.

Updates to particular `Merch` instances come from the sub-presentations.
Tests were added for the notification links between `InventoryPresentation`
and its `MerchPresentation`s.

Added a comparison function to `MerchPresentation` for sorting; it uses the
comparators defined on `Merch`.

In the course of implementing `InventoryPresentation`, I realized that
`add(_:)` had never been implemented in `MarketList` (!!), so that was added
along with tests.