Tests for `TripPresentation`
This commit largely parallels the previous one for `InventoryPresentation`.

Basic functionality for `TripPresentation` is established. The class exposes
a sorted list of `PurchasePresentation`s, and accepts additions of
`Purchases` and deletion by index.

It will optionally split the list of sub-presentations by their checked-off
state, sorting the two sublists by the master sort key.

Updates to particular `Purchase`s instances come from the sub-presentations.
Tests were added for the notification links between `TripPresentation` and
its `PurchasePresentation`s.

Worth noting in particular is that there is a separate notification for
modifications to the `Merch` fields of a sub-presentation, since the `Merch`
itself must be passed up so that the inventory can be updated.

Added a comparison function to `PurchasePresentation` for sorting; it uses
the comparators defined on `Purchase`.

Added some missing tests to `InventoryPresentation`.