Removed asking for purchase permission
`MerchPresentation` no longer has the `shouldPurchase()` event closure, and
does no test in the `purchase()` method before making changes.

After further consideration of this event route, there doesn't seem to be any
circumstance where any of the higher-level presentations would tell the
`MerchPresentation` that the purchase should not go forward, at least in the
simple, non-dialogue-based procedure. The `MerchPresentation` already will know
whether its item is on the trip: the decision to do purchase or remove will
have already been made.

It only makes sense for this action to be cancellable if the user goes through
the entry dialogue, to add a quantity or note.