Converted table views to use presentations
The table controller is now constructed with a `MarketPresentation`. It
takes the `InventoryPresentation` and uses its sub-presentations to
configure the cells.

Presentations added `NSObject` inheritance as necessary to be visible on
the ObjC side. The classes themselves and needed fields were made public.

Dummy data was moved out of the controllers and into extensions on the
presentations, in the file Presentations+Dummy.

SwiftLint flagged some whitespace errors in test files.